64-bit type changes and print format specifier cleanup; added suppressReplyNotifications to Calendar API; renamed newSubscriptionVideosFeedLink for happier static analysis; fixed double allocation in single-entry feeds

+++ trunk/Source/BaseClasses/GDataQuery.h
@@ -73,11 +73,11 @@ _EXTERN NSString* const kGDataQueryResultServiceDocument _INITIALIZE_AS(@"atom-s
 - (void)setFeedURL:(NSURL *)feedURL;
 
 // startIndex and maxResults use -1 to indicate parameter is not set
-- (int)startIndex;
-- (void)setStartIndex:(int)startIndex;
+- (NSInteger)startIndex;
+- (void)setStartIndex:(NSInteger)startIndex;
 
-- (int)maxResults;
-- (void)setMaxResults:(int)maxResults;
+- (NSInteger)maxResults;
+- (void)setMaxResults:(NSInteger)maxResults;
 
 - (NSString *)fullTextQueryString;
 - (void)setFullTextQueryString:(NSString *)str;
@@ -141,12 +141,12 @@ _EXTERN NSString* const kGDataQueryResultServiceDocument _INITIALIZE_AS(@"atom-s
 
 // if val==invalidVal, the parameter is removed
 - (void)addCustomParameterWithName:(NSString *)name
-                          intValue:(int)val
-                    removeForValue:(int)invalidVal;
+                          intValue:(NSInteger)val
+                    removeForValue:(NSInteger)invalidVal;
 
 // if the named parameter is not found, missingVal is returned
-- (int)intValueForParameterWithName:(NSString *)name
-              missingParameterValue:(int)missingVal;
+- (NSInteger)intValueForParameterWithName:(NSString *)name
+                    missingParameterValue:(NSInteger)missingVal;
 
 // convenience methods for boolean parameters
 - (void)addCustomParameterWithName:(NSString *)name
