+++ trunk/Source/GDataFramework.m
@@ -15,7 +15,7 @@
 
 #include "GDataFramework.h"
 
-void GDataFrameworkVersion(long* major, long* minor, long* release) {
+void GDataFrameworkVersion(NSUInteger* major, NSUInteger* minor, NSUInteger* release) {
   if (major)   *major = 1;
   if (minor)   *minor = 7;
   if (release) *release = 0;
